requirements.txt
fastapi
uvicorn
jinja2
firebase-admin
python-dotenv
fastapi-login 
python-multipart
gspread 
google-auth